test: add tests for use cases

+++ internal/usecase/implementation/productUsecase.go
@@ -14,9 +14,35 @@ func NewProductUseCase(productRepository repository.ProductRepository) *ProductU
 }
 
 func (uc *ProductUseCase) CreateProduct(inputProduct *domain.InputProduct) (*domain.Product, error) {
+	err := isValidType(inputProduct.Type)
+	if err != nil {
+		return nil, err
+	}
+	if inputProduct.PvzId == nil {
+		return nil, domain.ErrEmptyPvzID
+	}
 	return uc.productRepository.CreateProduct(inputProduct)
 }
 
 func (uc *ProductUseCase) DeleteLastProductForPVZ(pvzId *string) error {
+	if pvzId == nil || *pvzId == "" {
+		return domain.ErrEmptyPvzID
+	}
 	return uc.productRepository.DeleteLastProductForPVZ(pvzId)
+}
+
+func isValidType(productType *string) error {
+	if productType == nil {
+		return domain.ErrInvalidInputData
+	}
+	var allowedProductTypes = map[string]struct{}{
+		"электроника": {},
+		"одежда":      {},
+		"обувь":       {},
+	}
+	_, ok := allowedProductTypes[*productType]
+	if !ok {
+		return domain.ErrInvalidProductType
+	}
+	return nil
 }
